% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.IndicatorData.R
\name{plot.indicatorData}
\alias{plot.indicatorData}
\title{Plot Method for indicatorData Objects}
\usage{
\method{plot}{indicatorData}(x, indicatorList = NULL, topN = 3, facet = FALSE, ...)
}
\arguments{
\item{x}{An object of class `indicatorData`.}

\item{indicatorList}{Optional vector of indicator names to plot.}

\item{topN}{Number of top indicators to plot if no list is provided. Default is 3.}

\item{facet}{Logical; if TRUE, use facetting instead of overlaying lines. Default is FALSE.}

\item{...}{Additional graphical parameters (currently ignored).}
}
\description{
Plots average annual values for top N or specified indicators.
}
