% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.IndicatorData.R
\name{print.indicatorData}
\alias{print.indicatorData}
\title{Print Method for indicatorData Objects (With Indicator Filter)}
\usage{
\method{print}{indicatorData}(x, indicatorList = NULL, ...)
}
\arguments{
\item{x}{An object of class `indicatorData`.}

\item{indicatorList}{Optional character vector of indicator names to preview.
If NULL or empty, will display top 3 indicators.}

\item{...}{Ignored.}
}
\description{
Prints a summary line and a preview of selected or top indicators.
}
