% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIndicators.R
\name{readIndicators}
\alias{readIndicators}
\title{Read Indicator CSV File}
\usage{
readIndicators(filename, verbose = TRUE)
}
\arguments{
\item{filename}{Path to the `indicators_COUNTRYCODE.csv` file.}

\item{verbose}{Logical; if TRUE (default), prints summary and load time.}
}
\value{
An S3 object of class `indicatorData`.
}
\description{
Loads a country indicator CSV file, handles formatting quirks, assigns classes,
and returns an object of class `indicatorData`.
}
